requirements.txt

Flask==3.0.3
flatbuffers==23.5.26
langchain==0.1.15
langchain-community==0.0.32
langchain-core==0.1.52
langchain-openai==0.1.4
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
langcodes==3.4.0
langsmith==0.1.43
pinecone-client==3.2.2
streamlit==1.36.0
streamlit-chat==0.1.1
gunicorn==22.0.0
python-dotenv
sentence-transformers

